Refacto : moved functions that update mode pages & logical block pretection to mode.c / mode.h
