+++ include/mode.h
@@ -46,12 +46,17 @@ int add_mode_vendor_25h_mode_pages(struct lu_phy_attr *lu);
 int add_mode_encryption_mode_attribute(struct lu_phy_attr *lu);
 int add_mode_behavior_configuration(struct lu_phy_attr *lu);
 
-int	 add_mode_device_capabilities(struct lu_phy_attr *lu);
-int	 add_mode_transport_geometry(struct lu_phy_attr *lu);
-int	 add_mode_element_address_assignment(struct lu_phy_attr *lu);
-int	 update_prog_early_warning(struct lu_phy_attr *lu);
-int	 update_logical_block_protection(struct lu_phy_attr *lu, uint8_t *buf);
-void dealloc_all_mode_pages(struct lu_phy_attr *lu);
-int	 add_smc_mode_page_drive_configuration(struct lu_phy_attr *lu);
+int		add_mode_device_capabilities(struct lu_phy_attr *lu);
+int		add_mode_transport_geometry(struct lu_phy_attr *lu);
+int		add_mode_element_address_assignment(struct lu_phy_attr *lu);
+int		update_prog_early_warning(struct lu_phy_attr *lu);
+int		update_logical_block_protection(struct lu_phy_attr *lu, uint8_t *buf);
+uint8_t set_lbp(struct scsi_cmd *cmd, uint8_t *buf, int len);
+void	dealloc_all_mode_pages(struct lu_phy_attr *lu);
+int		add_smc_mode_page_drive_configuration(struct lu_phy_attr *lu);
+
+void	set_mode_compression(struct scsi_cmd *cmd, uint8_t *p);
+void	set_device_configuration(struct scsi_cmd *cmd, uint8_t *p);
+uint8_t set_device_configuration_extension(struct scsi_cmd *cmd, uint8_t *p);
 
 #endif /* MODE_H */
