+++ usr/mode.c
@@ -32,6 +32,8 @@
 #include "ssc.h"
 #include "smc.h"
 #include "be_byteshift.h"
+#include "mode.h"
+#include "mhvtl_log.h"
 
 static char *mode_rw_error_recover		  = "Read/Write Error Recovery";
 static char *mode_disconnect_reconnect	  = "Disconnect/Reconnect";
@@ -383,6 +385,42 @@ int add_mode_data_compression(struct lu_phy_attr *lu) {
 	return 0;
 }
 
+void set_mode_compression(struct scsi_cmd *cmd, uint8_t *p) {
+	struct lu_phy_attr				*lu		 = cmd->lu;
+	struct priv_lu_ssc				*lu_priv = lu->lu_private;
+	struct ssc_personality_template *pm		 = lu_priv->pm;
+
+	int dce = p[2] & 0x80;
+
+	MHVTL_DBG(2, " Data Compression Enable   : %s (0x%02x)",
+			  (p[2] & 0x80) ? "Yes" : "No", p[2]);
+	MHVTL_DBG(2, " Data Compression Capable  : %s",
+			  (p[2] & 0x40) ? "Yes" : "No");
+	MHVTL_DBG(2, " Data DeCompression Enable : %s (0x%02x)",
+			  (p[3] & 0x80) ? "Yes" : "No", p[3]);
+	MHVTL_DBG(2, " Compression Algorithm     : 0x%04x",
+			  get_unaligned_be32(&p[4]));
+	MHVTL_DBG(2, " DeCompression Algorithm   : 0x%04x",
+			  get_unaligned_be32(&p[8]));
+	MHVTL_DBG(2, " Report Exception on Decompression: 0x%02x",
+			  (p[3] & 0x6) >> 5);
+
+	if (dce) { /* Data Compression Enable bit set */
+		MHVTL_DBG(1, " Setting compression");
+		if (pm->set_compression) {
+			pm->set_compression(&lu->mode_pg,
+								lu_priv->configCompressionFactor);
+			set_lp11_compression(1); /* Update LogPage 11 compression bit */
+		}
+	} else {
+		MHVTL_DBG(1, " Clearing compression");
+		if (pm->clear_compression) {
+			pm->clear_compression(&lu->mode_pg);
+			set_lp11_compression(0); /* Update LogPage 11 compression bit */
+		}
+	}
+}
+
 /*
  * Device Configuration
  * SSC3-8.3.3
@@ -432,6 +470,35 @@ int add_mode_device_configuration(struct lu_phy_attr *lu) {
 	mp->description = mode_device_configuration;
 
 	return 0;
+}
+
+void set_device_configuration(struct scsi_cmd *cmd, uint8_t *p) {
+	struct lu_phy_attr				*lu		 = cmd->lu;
+	struct priv_lu_ssc				*lu_priv = cmd->lu->lu_private;
+	struct ssc_personality_template *pm		 = lu_priv->pm;
+
+	MHVTL_DBG(2, " Report Early Warning   : %s",
+			  (p[8] & 0x01) ? "Yes" : "No");
+	MHVTL_DBG(2, " Software Write Protect : %s",
+			  (p[10] & 0x04) ? "Yes" : "No");
+	MHVTL_DBG(2, " WORM Tamper Read Enable: %s",
+			  (p[15] & 0x80) ? "Yes" : "No");
+
+	MHVTL_DBG(2, " Setting device compression Algorithm");
+	if (p[14]) { /* Select Data Compression Alg */
+		MHVTL_DBG(2, "  Mode Select->Setting compression: %d", p[14]);
+		if (pm->set_compression) {
+			pm->set_compression(&lu->mode_pg,
+								lu_priv->configCompressionFactor);
+			set_lp11_compression(1); /* Update LogPage 11 compression bit */
+		}
+	} else {
+		MHVTL_DBG(2, "  Mode Select->Clearing compression");
+		if (pm->clear_compression) {
+			pm->clear_compression(&lu->mode_pg);
+			set_lp11_compression(0); /* Update LogPage 11 compression bit */
+		}
+	}
 }
 
 int add_mode_device_configuration_extension(struct lu_phy_attr *lu) {
@@ -485,6 +552,82 @@ int add_mode_device_configuration_extension(struct lu_phy_attr *lu) {
 	mp->description = mode_device_configuration_extension;
 
 	return 0;
+}
+
+uint8_t set_device_configuration_extension(struct scsi_cmd *cmd, uint8_t *p) {
+	struct lu_phy_attr				*lu		  = cmd->lu;
+	struct priv_lu_ssc				*lu_priv  = cmd->lu->lu_private;
+	struct ssc_personality_template *pm		  = lu_priv->pm;
+	uint8_t							*sam_stat = &cmd->dbuf_p->sam_stat;
+
+	struct mode *mp;
+	struct s_sd	 sd;
+	int			 page_code_len;
+	int			 write_mode;
+	int			 pews; /* Programable Early Warning Size */
+
+	mp = lookup_pcode(&lu->mode_pg, MODE_DEVICE_CONFIGURATION, 1);
+
+	/* Code error
+	 * Any device supporting this should have this mode page defined */
+	if (!mp) {
+		sam_hardware_error(E_INTERNAL_TARGET_FAILURE, sam_stat);
+		return SAM_STAT_CHECK_CONDITION;
+	}
+
+	page_code_len = get_unaligned_be16(&p[2]);
+
+	if (page_code_len != 0x1c) {
+		sd.byte0		 = SKSV;
+		sd.field_pointer = 2;
+		MHVTL_LOG("Unexpected page code length.. Unexpected results");
+		sam_illegal_request(E_INVALID_FIELD_IN_PARMS, &sd, sam_stat);
+		return SAM_STAT_CHECK_CONDITION;
+	}
+
+	write_mode = (p[5] & 0xf0) >> 4;
+	if (write_mode > 1) {
+		MHVTL_LOG("Unsupported write mode: 0x%x", write_mode);
+		sd.byte0		 = SKSV | BPV | 7; /* bit 7 */
+		sd.field_pointer = 5;
+		sam_illegal_request(E_INVALID_FIELD_IN_PARMS, &sd, sam_stat);
+		return SAM_STAT_CHECK_CONDITION;
+	}
+	MHVTL_DBG(2, "%s mode", write_mode ? "Append-only" : "Overwrite-allowed");
+
+	pews = get_unaligned_be16(&p[6]);
+	if (pm->drive_supports_prog_early_warning) {
+		MHVTL_DBG(2, "Set Programable Early Warning Size: %d", pews);
+		lu_priv->prog_early_warning_sz = pews;
+		update_prog_early_warning(lu);
+	} else {
+		MHVTL_DBG(2, "Programable Early Warning Size not supported"
+					 " by this device");
+	}
+
+	MHVTL_DBG(2, "Volume containing encrypted logical blocks "
+				 "requires encryption: %d",
+			  p[8] & 0x01);
+
+	if (pm->drive_supports_append_only_mode) {
+		/* Can't reset append-only mode via mode page ssc4 8.3.8 */
+		if (lu_priv->append_only_mode && write_mode == 0) {
+			MHVTL_LOG("Can't reset append only mode via mode page");
+			sam_illegal_request(E_INVALID_FIELD_IN_PARMS,
+								NULL, sam_stat);
+			return SAM_STAT_CHECK_CONDITION;
+		}
+		if (write_mode) {
+			lu_priv->append_only_mode = write_mode;
+			lu_priv->allow_overwrite  = FALSE;
+		}
+	}
+
+	/* Now update our copy of this mode page */
+	mp->pcodePointer[5] &= 0x0f;
+	mp->pcodePointer[5] |= write_mode << 4;
+
+	return SAM_STAT_GOOD;
 }
 
 int add_mode_medium_partition(struct lu_phy_attr *lu) {
@@ -1015,6 +1158,22 @@ int update_logical_block_protection(struct lu_phy_attr *lu, uint8_t *buf) {
 				  lu_priv->LBP_W ? "True" : "False");
 	}
 	return SAM_STAT_GOOD;
+}
+
+uint8_t set_lbp(struct scsi_cmd *cmd, uint8_t *buf, int len) {
+	struct priv_lu_ssc				*lu_priv  = cmd->lu->lu_private;
+	struct ssc_personality_template *pm		  = lu_priv->pm;
+	uint8_t							*sam_stat = &cmd->dbuf_p->sam_stat;
+	struct s_sd						 sd;
+
+	/* OK, the drive supports Logical Block Protection - good to go */
+	if (pm->drive_supports_LBP) {
+		return update_logical_block_protection(cmd->lu, buf);
+	}
+	sd.byte0		 = SKSV | CD;
+	sd.field_pointer = 1;
+	sam_illegal_request(E_INVALID_FIELD_IN_CDB, &sd, sam_stat);
+	return SAM_STAT_CHECK_CONDITION;
 }
 
 int add_smc_mode_page_drive_configuration(struct lu_phy_attr *lu) {
